misc/webapi-uiautomation-tests/html5-extra/attributes/anchor_href.htm

<!DOCTYPE HTML>
<head>
<title>HTMLAnchorElement href Attribute Test</title>
<meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
<link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
<link rel="help" href="http://dev.w3.org/html5/spec/text-level-semantics.html#the-a-element" />
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<pre>
HTMLAnchorElement Tests For The 'href' attribute
</pre>
<table id='testtable' border='1'>
<tr>
 <td>Test Result</td><td>Test Description</td><td>Test Value</td>
</tr>
<tr>
 <td id='test_1_result'></td>
 <td id='test_1_description'>Test get/set on a relative href attribute</td>
 <td id='test_1_value'>anchor.htm</td>
</tr>
<tr>
 <td id='test_2_result'></td>
 <td id='test_2_description'>Test get/set on a absolute href attribute</td>
 <td id='test_2_value'>http://somesite.com/anchor.htm</td>
</tr>
<tr>
 <td id='test_3_result'></td>
 <td id='test_3_description'>Test href attribute is an empty string</td>
 <td id='test_3_value'>empty string</td>
</tr>
<tr>
 <td id='test_4_result'></td>
 <td id='test_4_description'>Test for non-existent href attribute</td>
 <td id='test_4_value'>null</td>
</tr>
<tr>
 <td id='test_5_result'></td>
 <td id='test_5_description'>Test that a href attribute contains search string</td>
 <td id='test_5_value'>http://www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter</td>
</tr>
<tr>
 <td id='test_6_result'></td>
 <td id='test_6_description'>Test that a href attribute contains the username/password and the search string</td>
 <td id='test_6_value'>http://username:password@www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter</td>
</tr>
<tr>
 <td id='test_7_result'></td>
 <td id='test_7_description'>Test that a href attribute has not been lowercased</td>
 <td id='test_7_value'>http://WWW.SomeSite.com/INDEX.html</td>
</tr>
</table>
<pre>
http://dev.w3.org/html5/spec/text-level-semantics.html#the-a-element

interface HTMLAnchorElement : HTMLElement {
  stringifier attribute DOMString href;
           attribute DOMString target;
           attribute DOMString rel;
  readonly attribute DOMTokenList relList;
           attribute DOMString media;
           attribute DOMString hreflang;
           attribute DOMString type;
           attribute DOMString text;

  // URL decomposition IDL attributes
           attribute DOMString protocol;
           attribute DOMString host;
           attribute DOMString hostname;
           attribute DOMString port;
           attribute DOMString pathname;
           attribute DOMString search;
           attribute DOMString hash;
};
</pre>

<hr>
<div id='tags'>Tags Used For Testing</div>
<a id='anchor_1' href='anchor.htm'>Simple Anchor Relative Path</a><br>
<a id='anchor_2' href='http://somesite.com/anchor.htm'>Simple Anchor Non-Relative Path</a><br>
<a id='anchor_3' href=''>Empty String Anchor</a><br>
<a id='anchor_4'>Anchor with no attributes</a><br>
<a id='anchor_5' href='http://www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter'>href with a search string that is urlencoded</a><br>
<a id='anchor_6' href='http://username:password@www.anothersite.com/path/page.htm?parameter=this%20is%a%20parameter'>href with a username, password and search string that is urlencoded</a><br>
<a id='anchor_7' href='http://WWW.SomeSite.com/INDEX.html'>Simple Anchor with some uppercase letters</a><br>
<hr>

<div id="log"></div>
<script>

    var iTests = document.getElementById('testtable').rows.length;

    for (var iTest=1;iTest<iTests;iTest++)
    {
       test(function()
       {
        oAnchor     = document.getElementById('anchor_' + iTest);
        description = document.getElementById('test_' + iTest + '_description').innerHTML;
        value       = document.getElementById('test_' + iTest + '_value').innerHTML;

		// null needs to be special cased, same with an empty string
        if( value == 'null')
        {
            value = null;
        }
        else if( value == 'empty string')
        {
            value = '';
        }

        assert_equals(description, value, oAnchor.getAttribute('href'));

        oAnchor.setAttribute('href', '');
        assert_equals(description, '', oAnchor.getAttribute('href'));

		// remove the attribute for null
        if (value == null)
        {
            oAnchor.removeAttribute('href');
        }
        else
        {
            oAnchor.setAttribute('href', value);
        }
        assert_equals(description, value, oAnchor.getAttribute('href'));

        document.getElementById('test_'+ testnumber +'_result').innerHTML = 'Pass';
        }, 'anchor_' + iTest);
    }

</script>
</body>
</html>
